requirements.txt
asgiref==3.8.1
Django==5.0.6
filelock==3.14.0
fsspec==2024.3.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
sqlparse==0.5.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
